Add checks for node in aarch64 vector cost modeling
After removing STMT_VINFO_MEMORY_ACCESS_TYPE we now ICE when costing
for scalar stmts required in the epilog since the cost model tries
to pattern-match gathers (an earlier patch tried to improve this
by introducing stmt groups, but that was on hold due to negative
feedback).  The following shot-cuts those attempts when node is NULL
as that then cannot be a vector stmt.  Another possibility would be
to gate on vect_body, or restructure everything.

Note we now ensure that when m_costing_for_scalar node is NULL.

	* config/aarch64/aarch64.cc (aarch64_detect_vector_stmt_subtype):
	Check for node before dereferencing.
	(aarch64_vector_costs::add_stmt_cost): Likewise.