fix cris-elf build with binutils-2.45
Since the cris port was added to gcc it has passed --em=criself
to gas, as an abbreviation for --emulation=criself.  Starting with
binutils-2.45 that causes a hard error in gas due to ambiguity with
another option.

Fixed by replacing the abbreviation with the complete option.

Tested by building a cross to cris-elf with binutils-2.45, which
failed before but now succeeds.

gcc/
	PR target/121336
	* config/cris/cris.h: Do not abbreviate --emulation.

Signed-off-by: Mikael Pettersson <mikpelinux@gmail.com>