tree-optimization/121362 - missed FRE through aggregate copy
The following streamlines and generalizes how we find the common
base of the lookup ref and a kill ref when looking through
aggregate copies.  In particular this tries to deal with all
variants of punning that happens on the inner MEM_REF after
forwarding of address taken components of the common base.

	PR tree-optimization/121362
	* tree-ssa-sccvn.cc (vn_reference_lookup_3): Generalize
	aggregate copy handling.

	* gcc.dg/tree-ssa/ssa-fre-105.c: New testcase.
	* gcc.dg/tree-ssa/ssa-fre-106.c: Likewise.