Use x86 GFNI for vectorized constant byte shifts/rotates
The GFNI AVX gf2p8affineqb instruction can be used to implement
vectorized byte shifts or rotates. This patch uses them to implement
shift and rotate patterns to allow the vectorizer to use them.
Previously AVX couldn't do rotates (except with XOP) and had to handle
8 bit shifts with a half throughput 16 bit shift.

This is only implemented for constant shifts. In theory it could
be used with a lookup table for variable shifts, but it's unclear
if it's worth it.

The vectorizer cost model could be improved, but seems to work for now.
It doesn't model the true latencies of the instructions. Also it doesn't
account for the memory loading of the mask, assuming that for a loop
it will be loaded outside the loop.

The instructions would also support more complex patterns
(e.g. arbitary bit movement or inversions), so some of the tricks
applied to ternlog could be applied here too to collapse
more code. It's trickier because the input patterns
can be much longer since they can apply to every bit individually. I didn't
attempt any of this.

There's currently no test case for the masked/cond_ variants, they seem
to be difficult to trigger with the vectorizer. Suggestions for a test
case for them welcome.

gcc/ChangeLog:

	* config/i386/i386-expand.cc (ix86_vgf2p8affine_shift_matrix):
	New function to lookup shift/rotate matrixes for gf2p8affine.
	* config/i386/i386-protos.h (ix86_vgf2p8affine_shift_matrix):
	Declare new function.
	* config/i386/i386.cc (ix86_shift_rotate_cost): Add cost model
	for shift/rotate implemented using gf2p8affine.
	* config/i386/sse.md (VI1_AVX512_3264): New mode iterator.
	(<insn><mode>3): Add GFNI case for shift patterns.
	(cond_<insn><mode>3): New pattern.
	(<insn><mode>3<mask_name>): Dito.
	(<insn>v16qi): New rotate pattern to handle XOP V16QI case
	and GFNI.
	(rotl<mode>3, rotr<mode>3): Exclude V16QI case.

gcc/testsuite/ChangeLog:

	* gcc.target/i386/shift-gf2p8affine-1.c: New test
	* gcc.target/i386/shift-gf2p8affine-2.c: New test
	* gcc.target/i386/shift-gf2p8affine-3.c: New test
	* gcc.target/i386/shift-v16qi-4.c: New test
	* gcc.target/i386/shift-gf2p8affine-5.c: New test
	* gcc.target/i386/shift-gf2p8affine-6.c: New test
	* gcc.target/i386/shift-gf2p8affine-7.c: New test