aarch64: Improve svdupq_lane expension for big-endian [PR121293]
If the index to svdupq_lane is variable, or is outside the range of
the .Q form of DUP, the fallback expansion is to convert to VNx2DI and
use TBL.  The problem in this PR was that the conversion used subregs,
and on big-endian targets, a bitcast from VNx2DI to another element size
requires a REV[BHW] in the best case or a spill and reload in the worst
case.  (See the comment at the head of aarch64-sve.md for details.)

Here we want the conversion to act like svreinterpret, so it should
use aarch64_sve_reinterpret instead of subregs.

gcc/
	PR target/121293
	* config/aarch64/aarch64-sve-builtins-base.cc (svdupq_lane::expand):
	Use aarch64_sve_reinterpret instead of subregs.  Explicitly
	reinterpret the result back to the required mode, rather than
	leaving the caller to take a subreg.

gcc/testsuite/
	PR target/121293
	* gcc.target/aarch64/sve/acle/general/dupq_lane_9.c: New test.