aarch64: Drop unnecessary GPs in svcmp_wide PTEST patterns
Patterns that fuse a predicate operation P with a PTEST use
aarch64_sve_same_pred_for_ptest_p to test whether the governing
predicates of P and the PTEST are compatible.  Most patterns were also
written as define_insn_and_rewrites, with the rewrite replacing P's
original governing predicate with PTEST's.  This ensures that we don't,
for example, have both a .H PTRUE for the PTEST and a .B PTRUE for a
comparison that feeds the PTEST.

The svcmp_wide* patterns were missing this rewrite, meaning that we did
have redundant PTRUEs.

gcc/
	* config/aarch64/aarch64-sve.md
	(*aarch64_pred_cmp<cmp_op><mode>_wide_cc): Turn into a
	define_insn_and_rewrite and rewrite the governing predicate
	of the comparison so that it is identical to the PTEST's.
	(*aarch64_pred_cmp<cmp_op><mode>_wide_ptest): Likewise.

gcc/testsuite/
	* gcc.target/aarch64/sve/acle/general/cmpeq_1.c: Check the number
	of PTRUEs.
	* gcc.target/aarch64/sve/acle/general/cmpge_5.c: New test.
	* gcc.target/aarch64/sve/acle/general/cmpge_6.c: Likewise.
	* gcc.target/aarch64/sve/acle/general/cmpgt_5.c: Likewise.
	* gcc.target/aarch64/sve/acle/general/cmpgt_6.c: Likewise.
	* gcc.target/aarch64/sve/acle/general/cmple_5.c: Likewise.
	* gcc.target/aarch64/sve/acle/general/cmple_6.c: Likewise.
	* gcc.target/aarch64/sve/acle/general/cmplt_5.c: Likewise.
	* gcc.target/aarch64/sve/acle/general/cmplt_6.c: Likewise.
	* gcc.target/aarch64/sve/acle/general/cmpne_3.c: Likewise.