Fail early when SLP instance build fails
The following makes us fail earlier when parts of the SLP build fails.
Currently we rely on hybrid stmt detection later to discover not all
stmts are covered by SLP, but this code should go away.  I've also
seen a case of a missed gcond SLP build that went undetected.  So
the following makes us fail during vect_analyze_slp if any of the
SLP instances we expect to discover fails.

	* tree-vect-slp.cc (vect_analyze_slp): When analyzing a loop
	and slp instance discovery fails, immediately fail the whole
	process.