fortran: Remove span overwrite with pointer assignments
Remove an overwrite of the array descriptor span field when pointer-
assigning from a polymorphic function result to a non-polymorphic
pointer.  That overwrite doesn't make sense because the span is
determined by the memory layout of the array; we can't change it
without also changing the data pointer.

gcc/fortran/ChangeLog:

	* trans-expr.cc (gfc_trans_pointer_assignment): Remove overwrite
	of the span after assignment of the array descriptor in the
	polymorphic function result to non-polymorphic pointer case.

gcc/testsuite/ChangeLog:

	* gfortran.dg/pointer_assign_16.f90: New test.